+++ src/webgpu/api/operation/resource_init/texture_zero.spec.ts
@@ -1,4 +1,11 @@
-export const description = 'Test uninitialized textures are initialized to zero when read.';
+export const description = `
+Test uninitialized textures are initialized to zero when read.
+
+TODO:
+- 1d, 3d
+- test by sampling depth/stencil
+- test by copying out of stencil
+`;
 
 // TODO: This is a test file, it probably shouldn't export anything.
 // Everything that's exported should be moved to another file.
