+++ src/webgpu/api/validation/render_pass/resolve.spec.ts
@@ -1,28 +1,6 @@
-export const description = `API Validation Tests for RenderPass Resolve.
-
-  Test Coverage:
-    - When resolveTarget is not null:
-      - Test that the colorAttachment is multisampled:
-        - A single sampled colorAttachment should generate an error.
-      - Test that the resolveTarget is single sampled:
-        - A multisampled resolveTarget should generate an error.
-      - Test that the resolveTarget has usage OUTPUT_ATTACHMENT:
-        - A resolveTarget without usage OUTPUT_ATTACHMENT should generate an error.
-      - Test that the resolveTarget's texture view describes a single subresource:
-        - A resolveTarget texture view with base mip {0, base mip > 0} and mip count of 1 should be
-          valid.
-          - An error should be generated when the resolve target view mip count is not 1 and base
-            mip is {0, base mip > 0}.
-        - A resolveTarget texture view with base array layer {0, base array layer > 0} and array
-          layer count of 1 should be valid.
-          - An error should be generated when the resolve target view array layer count is not 1 and
-            base array layer is {0, base array layer > 0}.
-      - Test that the resolveTarget's format is the same as the colorAttachment:
-        - An error should be generated when the resolveTarget's format does not match the
-          colorAttachment's format.
-      - Test that the resolveTarget's size is the same the colorAttachment:
-        - An error should be generated when the resolveTarget's height or width are not equal to
-          the colorAttachment's height or width.`;
+export const description = `
+Validation tests for render pass resolve.
+`;
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { GPUConst } from '../../../constants.js';
@@ -33,6 +11,25 @@ const kNumColorAttachments = 4;
 export const g = makeTestGroup(ValidationTest);
 
 g.test('resolve_attachment')
+  .desc(
+    `
+Test various validation behaviors when a resolveTarget is provided.
+
+- base case (valid).
+- resolve source is not multisampled.
+- resolve target is not single sampled.
+- resolve target missing OUTPUT_ATTACHMENT usage.
+- resolve target must have exactly one subresource:
+    - base mip level {0, >0}, mip level count {1, >1}.
+    - base array layer {0, >0}, array layer count {1, >1}.
+    - TODO: test zero subresources
+- resolve source and target have different formats.
+    - rgba8unorm -> {bgra8unorm, rgba8unorm-srgb}
+    - {bgra8unorm, rgba8unorm-srgb} -> rgba8unorm
+    - test with other color attachments having a different format
+- resolve source and target have different sizes.
+`
+  )
   .params([
     // control case should be valid
     { _valid: true },
